memdb: exclude random tests when race is enabled (#16478)
